+++ Sheenam.Api.Tests.Unit/Services/Foundations/Owners/OwnerServiceTests.Exceptions.RetrieveById.cs
@@ -7,8 +7,8 @@ using FluentAssertions;
 using Microsoft.Data.SqlClient;
 using Microsoft.Extensions.Hosting;
 using Moq;
-using Sheenam.Api.Models.Foundations.Owner;
-using Sheenam.Api.Models.Foundations.Owner.Exceptions;
+using Sheenam.Api.Models.Foundations.Owners;
+using Sheenam.Api.Models.Foundations.Owners.Exceptions;
 using System;
 using System.Threading.Tasks;
 using Xunit;
