+++ Sheenam.Api.Tests.Unit/Services/Foundations/Owners/OwnerServiceTests.Logic.Modify.cs
@@ -7,7 +7,7 @@ using FluentAssertions;
 using Force.DeepCloner;
 using Microsoft.Extensions.Hosting;
 using Moq;
-using Sheenam.Api.Models.Foundations.Owner;
+using Sheenam.Api.Models.Foundations.Owners;
 using System;
 using System.Linq;
 using System.Threading.Tasks;
