+++ Sheenam.Api.Tests.Unit/Services/Foundations/Owners/OwnerServiceTests.Logic.RetrieveAll.cs
@@ -5,7 +5,7 @@
 
 using FluentAssertions;
 using Moq;
-using Sheenam.Api.Models.Foundations.Owner;
+using Sheenam.Api.Models.Foundations.Owners;
 using System.Linq;
 using Xunit;
