+++ Sheenam.Api.Tests.Unit/Services/Foundations/Owners/OwnerServiceTests.Validations.RetrieveById.cs
@@ -8,8 +8,8 @@ using Moq;
 using System.Threading.Tasks;
 using System;
 using Xunit;
-using Sheenam.Api.Models.Foundations.Owner.Exceptions;
-using Sheenam.Api.Models.Foundations.Owner;
+using Sheenam.Api.Models.Foundations.Owners.Exceptions;
+using Sheenam.Api.Models.Foundations.Owners;
 using FluentAssertions;
 
 namespace Sheenam.Api.Tests.Unit.Services.Foundations.Owners
