+++ Sheenam.Api/Models/Foundations/Owners/Exceptions/FailedOwnerServiceException.cs
@@ -6,7 +6,7 @@
 using System;
 using Xeptions;
 
-namespace Sheenam.Api.Models.Foundations.Owner.Exceptions
+namespace Sheenam.Api.Models.Foundations.Owners.Exceptions
 {
     public class FailedOwnerServiceException : Xeption
     {
