+++ Sheenam.Api/Models/Foundations/Owners/Exceptions/FailedOwnerStorageException.cs
@@ -8,7 +8,7 @@
 using System;
 using Xeptions;
 
-namespace Sheenam.Api.Models.Foundations.Owner.Exceptions
+namespace Sheenam.Api.Models.Foundations.Owners.Exceptions
 {
     public class FailedOwnerStorageException : Xeption
     {
