+++ Sheenam.Api/Models/Foundations/Owners/Exceptions/OwnerServiceException.cs
@@ -5,7 +5,7 @@
 
 using Xeptions;
 
-namespace Sheenam.Api.Models.Foundations.Owner.Exceptions
+namespace Sheenam.Api.Models.Foundations.Owners.Exceptions
 {
     public class OwnerServiceException : Xeption
     {
