+++ Sheenam.Api/Models/Foundations/Owners/Exceptions/OwnerValidationException.cs
@@ -6,7 +6,7 @@
 
 using Xeptions;
 
-namespace Sheenam.Api.Models.Foundations.Owner.Exceptions
+namespace Sheenam.Api.Models.Foundations.Owners.Exceptions
 {
     public class OwnerValidationException : Xeption
     {
