+++ Sheenam.Api/Services/Foundations/Owners/OwnerService.Validations.cs
@@ -6,8 +6,8 @@
 using System;
 using System.Data;
 using Microsoft.Extensions.Hosting;
-using Sheenam.Api.Models.Foundations.Owner;
-using Sheenam.Api.Models.Foundations.Owner.Exceptions;
+using Sheenam.Api.Models.Foundations.Owners;
+using Sheenam.Api.Models.Foundations.Owners.Exceptions;
 using static System.Runtime.InteropServices.JavaScript.JSType;
 
 namespace Sheenam.Api.Services.Foundations.Owners
